backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Client
requests==2.31.0

# Redis Cache
redis==5.0.1

# Configuration Management
PyYAML==6.0.1

# Task Scheduling
APScheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==6.1.0
black==23.12.1

# WSGI Server (Production)
gunicorn==21.2.0
